fix: quota SUM numeric cast + diagnose docker exit + upload error
- quota.rs: cast SUM(bytes)::bigint so PostgreSQL returns BIGINT, not
  NUMERIC. sqlx's (i64, i64) decode rejects NUMERIC, which made every
  upload fail with HTTP 500 (the query previously died at text = uuid
  before reaching SUM, so this only surfaced after the migration fix).
- images handler: tracing::error the ServiceError before mapping to 500,
  so CI logs show the cause instead of a bare 'internal server error'.
- ci docker-build: add a 5s foreground startup probe to capture stdout/
  stderr before the detached run. The container exits with code 0 and
  empty logs, so the foreground probe is needed to see what happens.

+++ crates/service/src/quota.rs
@@ -50,7 +50,7 @@ impl QuotaService {
                     SELECT
                         COALESCE((SELECT max_bytes FROM quotas WHERE user_id = $1), $2::bigint),
                         COALESCE(
-                            (SELECT SUM(bytes) FROM images WHERE owner_id = $1 AND status != 'deleted'),
+                            (SELECT SUM(bytes)::bigint FROM images WHERE owner_id = $1 AND status != 'deleted'),
                             0
                         )
                     ",
